+++ src/main/java/io/hpb/web3/tx/FastRawTransactionManager.java
@@ -5,8 +5,6 @@ import java.math.BigInteger;
 
 import io.hpb.web3.crypto.Credentials;
 import io.hpb.web3.protocol.Web3;
-import io.hpb.web3.tx.response.Callback;
-import io.hpb.web3.tx.response.QueuingTransactionReceiptProcessor;
 import io.hpb.web3.tx.response.TransactionReceiptProcessor;
 
 
@@ -14,7 +12,7 @@ public class FastRawTransactionManager extends RawTransactionManager {
 
     private volatile BigInteger nonce = BigInteger.valueOf(-1);
 
-    public FastRawTransactionManager(Web3 web3, Credentials credentials, byte chainId) {
+    public FastRawTransactionManager(Web3 web3, Credentials credentials, int chainId) {
         super(web3, credentials, chainId);
     }
 
@@ -29,7 +27,7 @@ public class FastRawTransactionManager extends RawTransactionManager {
     }
 
     public FastRawTransactionManager(
-            Web3 web3, Credentials credentials, byte chainId,
+            Web3 web3, Credentials credentials, int chainId,
             TransactionReceiptProcessor transactionReceiptProcessor) {
         super(web3, credentials, chainId, transactionReceiptProcessor);
     }
